+++ localization/ndt_scan_matcher/src/ndt_scan_matcher_core.cpp
@@ -16,9 +16,9 @@
 
 #include "localization_util/matrix_type.hpp"
 #include "localization_util/pose_array_interpolator.hpp"
-#include "localization_util/tree_structured_parzen_estimator.hpp"
 #include "localization_util/util_func.hpp"
 #include "ndt_scan_matcher/particle.hpp"
+#include "tree_structured_parzen_estimator/tree_structured_parzen_estimator.hpp"
 
 #include <tier4_autoware_utils/geometry/geometry.hpp>
 #include <tier4_autoware_utils/transform/transforms.hpp>
