+++ localization/tree_structured_parzen_estimator/README.md
@@ -1,0 +1,5 @@
+# tree_structured_parzen_estimator
+
+`tree_structured_parzen_estimator`` is a package for black-box optimization.
+
+This package does not have a node, it is just a library.
